requirements.txt
streamlit
torch>=1.7,<2.1
Pillow
opencv-python